Update gradle and JAR files
Update all gradle files to force a clean before build.  This gets around
a problem we had with making our "fat" JAR files; namely, if there was a
change in a dependency (in a CTlib file, for instance) but no change in
any source files of the current project, then the JAR file wasn't being
rebuilt.  To get around this, whenever a user asks to do a "build" we
force a "clean" to be executed first.