Restructure build.gradle files and add Javadoc generation
1. Restructure build.gradle files for both Java and Android applications
using Gradle "configuration injection", that is we include common logic
for all subprojects in the "subprojects" closure of the parent project's
build.gradle file.  For example, for Java projects, the logic to copy
the output JAR file to the common "Distribute" folder is now in
CloudTurbine/build.gradle's "subprojects" closure.

2. Projects generate Javadoc and have the HTML files put into the common
JavaDoc folder; separate folders under JavaDoc for each project, for
example JavaDoc/CTlib.